backend/core-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
sqlalchemy
pyodbc
python-multipart